+++ geotrek/tourism/templates/tourism/sql/post_90_defaults.sql
@@ -27,6 +27,7 @@ ALTER TABLE tourism_informationdesk ALTER COLUMN accessibility SET DEFAULT '';
 -- geom
 -- eid
 ALTER TABLE tourism_informationdesk ALTER COLUMN uuid SET DEFAULT gen_random_uuid();
+ALTER TABLE tourism_informationdesk ALTER COLUMN provider SET DEFAULT '';
 
 
 -- TouristicContentCategory
@@ -77,6 +78,7 @@ ALTER TABLE tourism_touristiccontent ALTER COLUMN published SET DEFAULT FALSE;
 -- structure
 ALTER TABLE tourism_touristiccontent ALTER COLUMN date_insert SET DEFAULT now();
 ALTER TABLE tourism_touristiccontent ALTER COLUMN date_update SET DEFAULT now();
+ALTER TABLE tourism_touristiccontent ALTER COLUMN provider SET DEFAULT '';
 -- deleted
 
 
@@ -121,3 +123,4 @@ ALTER TABLE tourism_touristicevent ALTER COLUMN published SET DEFAULT FALSE;
 ALTER TABLE tourism_touristiccontent ALTER COLUMN date_insert SET DEFAULT now();
 ALTER TABLE tourism_touristiccontent ALTER COLUMN date_update SET DEFAULT now();
 -- deleted
+ALTER TABLE tourism_touristicevent ALTER COLUMN provider SET DEFAULT '';